Move code from site.pp into a profile
Prior to this commit, the code to setup bitbucket server was
contained directly in site.pp for specific use in this vagrant
stack.

After this commit, the code has been moved into a seperate repo
where it can be used more generally.

This commit removes the code and instead clones down the new repo
and includes the new profile module.